@@ -18,10 +18,10 @@
   },
   "homepage": "https://github.com/msooridev/react-skeleton#readme",
   "dependencies": {
-    "babel-preset-react": "^6.1.18",
-    "babelify": "^7.2.0",
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3",
-    "watchify": "^3.6.1"
+    "babel-preset-react": "^6.24.1",
+    "babelify": "^7.3.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "watchify": "^3.11.1"
   }
 }
