@@ -1,13 +1,14 @@
 var React = require('react');
 
-var ListItem = React.createClass({
-   render: function () {
-       return (
-           <li>
-               <h4>{this.props.ingrediant}</h4>
-           </li>
-       );
-   } 
-});
+class ListItem extends React.Component {
+    render() {
+        return (
+            <li>
+                <h4>{this.props.ingrediant}</h4>
+            </li>
+        );
+    }
+}
 
 module.exports = ListItem;
+
